Add files via upload
This version lets you play with a friend. It has rule-checking and knows en-passent and castle. It makes a sound when you want to make an illegal move. If you resize the window, the squares will be resized too. There is no proper testing for checks and no pawn-promotion yet. However there is already a method, that can check, if a square is attacked.